pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tz_spotware"
version = "0.1.0"
description = "AI Engineer Take-Home: RAG application with LangChain and LangGraph"
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "langchain>=0.3.0,<0.4.0",
    "langgraph>=0.6.0,<0.7.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-openai>=0.2.0", # For ChatOpenAI (replaces deprecated langchain_community.chat_models)
    "sentence-transformers>=2.3.0",
    "faiss-cpu>=1.7.4",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0", # For OpenRouter compatibility
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6", # For file uploads
    "loguru>=0.7.0", # Structured logging
    "pydantic-settings>=2.12.0",
    "pydantic>=2.12.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 100
target-version = ['py312']

